LayoutTests/fast/canvas/webgl/array-set-invalid-arguments-expected.txt

Verifies that attempting to set invalid elements to a Typed Array throws an exception

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typedArray.set() threw exception TypeError: Not enough arguments.
PASS typedArray.set('hello world') threw exception TypeError: Invalid argument.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Index is out of range..
PASS typedArray.set() threw exception TypeError: Not enough arguments.
PASS typedArray.set('hello world') threw exception TypeError: Invalid argument.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Index is out of range..
PASS typedArray.set() threw exception TypeError: Not enough arguments.
PASS typedArray.set('hello world') threw exception TypeError: Invalid argument.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Index is out of range..
PASS typedArray.set() threw exception TypeError: Not enough arguments.
PASS typedArray.set('hello world') threw exception TypeError: Invalid argument.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Index is out of range..
PASS typedArray.set() threw exception TypeError: Not enough arguments.
PASS typedArray.set('hello world') threw exception TypeError: Invalid argument.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Index is out of range..
PASS typedArray.set() threw exception TypeError: Not enough arguments.
PASS typedArray.set('hello world') threw exception TypeError: Invalid argument.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Index is out of range..
PASS typedArray.set() threw exception TypeError: Not enough arguments.
PASS typedArray.set('hello world') threw exception TypeError: Invalid argument.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Index is out of range..
PASS successfullyParsed is true

TEST COMPLETE
